refactor: 가로 스크롤 커스텀 훅으로 분리
